Fix broken getAttributeSpellingListIndex for pragma attributes
We noticed when implementing a new pragma that the TableGen-generated function 
getAttributeSpellingListIndex() did not work for pragma attributes. It relies 
on the values in the enum AttributeList::Syntax and a new value 
AS_ContextSensitiveKeyword was added changing the value for AS_Pragma. 
Apparently no tests failed since no pragmas currently make use of the 
generated function.

To fix this we can move AS_Pragma back to the value that TableGen code expects. 
Also to prevent changes in the enum from breaking that routine again I added 
calls to getAttributeSpellingListIndex() in the unroll pragma code. That will 
cause some lit test failures if the order is changed. I added a comment to 
remind of this issue in the future.

This assumes we don’t need/want full TableGen support for 
AS_ContextSensitiveKeyword. It currently only appears in getAttrKind and no 
other TableGen-generated routines.

Patch by: mikerice

Differential Revision: https://reviews.llvm.org/D36473


git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@310483 91177308-0d34-0410-b5e6-96231b3b80d8